refactor(core): use usize for bake_scene resolution
Drop the usize -> u32 -> usize round-trip on every native bake.
The core function now takes `res: usize` directly, matching
`Volume::res`. The wasm shim (`wasm::bake_scene`) keeps its
`res: u32` JS-facing ABI and casts once at the boundary, so the
JS-side signature is unchanged.

The original issue 01 spec said `u32`, but that was written before
the WASM/native split made the cast obviously redundant for the
core function. The wasm-bindgen ABI is the only thing that needs
`u32`, and `wasm::bake_scene` already takes it.

+++ crates/atom-core/src/wasm.rs
@@ -65,7 +65,7 @@ impl BakeResult {
 #[wasm_bindgen]
 pub fn bake_scene(n: u32, l: u32, m: i32, res: u32) -> BakeResult {
     let scene = Scene::single_hydrogen(Orbital { n, l, m });
-    let v = volume::bake_scene(&scene, res);
+    let v = volume::bake_scene(&scene, res as usize);
     BakeResult {
         data: v.data,
         half_extent: v.half_extent as f32,
